Break sharesec into separate file
